Updates and folder removal done
