--- certificate-manager/src/index.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-
-const App: React.FC = () => {
-  return <h1>Hey there champ!! How are you doing?</h1>;
-};
-
-ReactDOM.render(<App />, document.getElementById('root'));
