+++ package.json
@@ -1,0 +1,42 @@
+{
+  "name": "certificate-manager",
+  "version": "1.0.0",
+  "description": "A certificate management application built with React and TypeScript",
+  "main": "index.js",
+  "scripts": {
+    "start": "webpack serve --open",
+    "dev": "webpack serve --mode development",
+    "build": "webpack --mode production",
+    "type-check": "tsc --noEmit",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": ["certificate", "management", "react", "typescript"],
+  "author": "Your Name",
+  "license": "ISC",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/KelliaKamikazi/certificate-manager-kellia"
+  },
+  "dependencies": {
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
+  },
+  "devDependencies": {
+    "@babel/core": "7.23.7",
+    "@babel/preset-env": "7.23.7",
+    "@babel/preset-react": "7.23.7",
+    "@babel/preset-typescript": "7.23.7",
+    "@types/node": "20.11.0",
+    "@types/react": "18.2.48",
+    "@types/react-dom": "18.2.18",
+    "babel-loader": "9.1.3",
+    "clean-webpack-plugin": "4.0.0",
+    "html-webpack-plugin": "5.6.0",
+    "source-map-loader": "5.0.0",
+    "ts-loader": "9.5.1",
+    "typescript": "5.3.3",
+    "webpack": "5.89.0",
+    "webpack-cli": "5.1.4",
+    "webpack-dev-server": "4.15.1"
+  }
+}
