+++ src/index.css
@@ -1,0 +1,14 @@
+body {
+    font-family: Arial, sans-serif;
+    margin: 0;
+    padding: 0;
+    background-color: #f0f0f0;
+  }
+  
+  h1 {
+    color: blue;
+    font-size: 24px;
+    text-align: center;
+    margin-top: 20px;
+  }
+
