+++ src/index.html
@@ -7,5 +7,6 @@
 </head>
 <body>
   <div id="root"></div>
+  <script src="bundle.js"></script>
 </body>
 </html>
